desloppify: split rimap-audit/src/writer/mod.rs into focused submodules
Sprint B item 3a. Mechanical decomposition; no behavior change.

Production-code portion of mod.rs went from 1244 LOC to 273 LOC
(~78% reduction). Tests stay in mod.rs (cross-cut emit/log/rotation).

  writer/mod.rs (was 1244 LOC, now 273 prod + 574 tests)
    - AuditOptions, AuditWriter, Inner, FailureInjection structs
    - open() ctor (acquires LOCK_EX, sets 0600)
    - path/process_id/rotate_bytes/bytes_written/on_disk_len accessors
    - suppressed_failures + force_next_write_failure (test-injection)
    - set_file_mode_0600 / set_parent_mode_0700 platform helpers
    - All existing unit tests

  writer/emit.rs (NEW, 205 LOC)
    - allocate_seq + lock_inner + emit (allocate-build-write skeleton)
    - write_record / write_record_inner (rotation-aware, fail_open
      suppression policy, fsync gating)
    - write_under_lock + needs_fsync free helpers

  writer/log.rs (NEW, 224 LOC)
    - log_auth, log_tool_start, log_tool_end, log_process_start,
      log_process_end family
    - ToolStartInputs, ToolEndInputs, ProcessStartInputs structs
    - From<Inputs> for record::* impls
    - Module-level rustdoc pinning the log_* family input convention

Visibility: lock_inner and emit promoted to pub(super) so log.rs can
call them; the rest stays private.
